Update: Fix bug in socket disconnect, add additional signal functions, and some clean up.
The socket close enumerations are being directly passed to shutdown().
This is not correct because they do not directly map.

Use the size_read and size_write already provided in the socket file.
This allows for the length to be exclusively a write (better practice).

Provide f_signal_wait() and f_signal_wait_until() functions that handle sigwaitinfo() and sigtimedwait() respectively.

Use uint8_t rather than unsigned short.
The fact that there is still a short in use here makes it clear that I have not even glanced at this file in a long long time.

+++ level_0/f_socket/c/socket.h
@@ -287,7 +287,7 @@ extern "C" {
  * @see shutdown()
  */
 #ifndef _di_f_socket_disconnect_
-  extern f_status_t f_socket_disconnect(f_socket_t * const socket, const unsigned short action);
+  extern f_status_t f_socket_disconnect(f_socket_t * const socket, const uint8_t action);
 #endif // _di_f_socket_disconnect_
 
 /**
@@ -388,6 +388,7 @@ extern "C" {
  * @param socket
  *   The socket structure.
  *   The socket.id must represent a valid socket file descriptor.
+ *   The socket.size_read is used to represent the buffer size in buffer and must not be larger than the actual size of the buffer.
  * @param flags
  *   Read flags.
  * @param buffer
@@ -474,6 +475,7 @@ extern "C" {
  * @param socket
  *   The socket structure.
  *   The socket.id must represent a valid socket file descriptor.
+ *   The socket.size_write is used to represent the buffer size in buffer and must not be larger than the actual size of the buffer.
  * @param flags
  *   Read flags.
  * @param buffer
